error refactoring
