+++ Sources/AmazonLocationiOSAuthSDK/Types/ApiTypes/ResponseTypes/AmazonBaseErrorResponse.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+public protocol AmazonBaseErrorResponse: Codable {
+    var errorCode: String { get }
+    var errorMessage: String { get }
+}
